Change account edit feature to theme 2
