+++ feature/account/edit/src/main/kotlin/app/k9mail/feature/account/edit/ui/server/settings/save/SaveServerSettingsContent.kt
@@ -7,12 +7,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.molecule.ContentLoadingErrorView
 import app.k9mail.core.ui.compose.designsystem.molecule.ErrorView
 import app.k9mail.core.ui.compose.designsystem.molecule.LoadingView
 import app.k9mail.core.ui.compose.designsystem.template.ResponsiveWidthContainer
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 import app.k9mail.feature.account.common.ui.loadingerror.rememberContentLoadingErrorViewState
 import app.k9mail.feature.account.edit.R
 
@@ -46,51 +44,6 @@ fun SaveServerSettingsContent(
                 )
             },
             modifier = Modifier.fillMaxSize(),
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun UpdateServerSettingsContentPreview() {
-    PreviewWithThemes {
-        SaveServerSettingsContent(
-            state = SaveServerSettingsContract.State(
-                isLoading = false,
-                error = null,
-            ),
-            contentPadding = PaddingValues(),
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun UpdateServerSettingsContentLoadingPreview() {
-    PreviewWithThemes {
-        SaveServerSettingsContent(
-            state = SaveServerSettingsContract.State(
-                isLoading = true,
-                error = null,
-            ),
-
-            contentPadding = PaddingValues(),
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun UpdateServerSettingsContentErrorPreview() {
-    PreviewWithThemes {
-        SaveServerSettingsContent(
-            state = SaveServerSettingsContract.State(
-                isLoading = false,
-                error = SaveServerSettingsContract.Failure.SaveServerSettingsFailed(
-                    message = "Error",
-                ),
-            ),
-            contentPadding = PaddingValues(),
         )
     }
 }
