+++ feature/account/edit/src/main/kotlin/app/k9mail/feature/account/edit/ui/server/settings/save/SaveServerSettingsScreen.kt
@@ -4,18 +4,14 @@ import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.common.mvi.observe
 import app.k9mail.core.ui.compose.designsystem.organism.TopAppBarWithBackButton
 import app.k9mail.core.ui.compose.designsystem.template.Scaffold
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.ThunderbirdTheme
 import app.k9mail.feature.account.common.ui.WizardNavigationBar
 import app.k9mail.feature.account.common.ui.WizardNavigationBarState
 import app.k9mail.feature.account.edit.ui.server.settings.save.SaveServerSettingsContract.Effect
 import app.k9mail.feature.account.edit.ui.server.settings.save.SaveServerSettingsContract.Event
 import app.k9mail.feature.account.edit.ui.server.settings.save.SaveServerSettingsContract.ViewModel
-import app.k9mail.feature.account.edit.ui.server.settings.save.fake.FakeSaveServerSettingsViewModel
 
 @Composable
 fun SaveServerSettingsScreen(
@@ -66,36 +62,6 @@ fun SaveServerSettingsScreen(
         SaveServerSettingsContent(
             state = state.value,
             contentPadding = innerPadding,
-        )
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-internal fun SaveServerSettingsScreenK9Preview() {
-    K9Theme {
-        SaveServerSettingsScreen(
-            title = "Incoming server settings",
-            onNext = {},
-            onBack = {},
-            viewModel = FakeSaveServerSettingsViewModel(
-                isIncoming = true,
-            ),
-        )
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-internal fun SaveServerSettingsScreenThunderbirdPreview() {
-    ThunderbirdTheme {
-        SaveServerSettingsScreen(
-            title = "Incoming server settings",
-            onNext = {},
-            onBack = {},
-            viewModel = FakeSaveServerSettingsViewModel(
-                isIncoming = true,
-            ),
         )
     }
 }
